requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.4
django-allauth==65.1.0
django-cloudinary-storage==0.3.0
django-contrib-comments==2.2.0
django-js-asset==2.2.0
django-mptt==0.16.0
dnspython==2.6.1
email_validator==2.1.2
filebrowser-safe==1.1.1
Flask==3.0.3
Flask-WTF==1.2.1
git-filter-repo==2.38.0
grappelli-safe==1.1.1
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Mezzanine==6.0.0
oauthlib==3.2.2
packaging==24.2
pillow==10.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stripe==11.2.0
tinycss2==1.2.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.3
whitenoise==6.8.2
WTForms==3.1.2
